Install passlib
Ansible 2.17 shows that password_hash filter is going to be deprecated
since Ansible 2.19.
I noticed this because passlib, a requirement for password_hash, is not
pulled when installing ansible-2.17 with pip.
Apparently, passlib is not being supported, three years without commits.
More info in ansible/ansible#81949 and in ansible/ansible#82652.
We can use some cli like mkpasswd instead.